backend/requirements.txt
absl-py==2.3.0
addict==2.4.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
apyori==1.1.2
asttokens==3.0.0
astunparse==1.6.3
backports.tarfile==1.2.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==6.0.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
docutils==0.21.2
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filestack-python==4.0.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.1
fpdf==1.7.2
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
justpy==0.14.0
keras==3.10.0
keyring==25.6.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.2.21
nltk==3.9.1
numpy==2.1.3
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
premailer==3.10.0
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
selectorlib==0.16.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
trafaret==2.0.2
traitlets==5.14.3
twine==6.1.0
types-psycopg2==2.9.21.20251012
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.22.1
w3lib==2.3.1
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xgboost==3.0.2
yagmail==0.15.293
zipp==3.22.0
python-jose[cryptography]